+++ src/modules/Recipes.lua
@@ -260,6 +260,8 @@ function Self:GetRecipeInfo(recipeOrOrder, isRecraft)
     end
 end
 
+local n = 0
+
 ---@param recipe CraftingRecipeSchematic
 ---@param reagents CraftingReagentInfo[]
 ---@param orderOrRecraftGUID? CraftingOrderInfo | string
@@ -267,8 +269,35 @@ end
 function Self:GetOperationInfo(recipe, reagents, orderOrRecraftGUID, applyConcentration)
     if not applyConcentration then applyConcentration = false end
 
+    -- Create operation info for recipe's that don't have any
+    if not recipe.hasCraftingOperationInfo then
+        local profInfo = C_TradeSkillUI.GetProfessionInfoByRecipeID(recipe.recipeID)
+        return  { recipeID = recipe.recipeID, baseSkill = profInfo.skillLevel, bonusSkill = profInfo.skillModifier, baseDifficulty = 0, bonusDifficulty = 1, isQualityCraft = false, quality = 1, craftingQuality = 1, craftingQualityID = 0, craftingDataID = 0, lowerSkillThreshold = 0, upperSkillTreshold = 0, guaranteedCraftingQualityID = 0, bonusStats = {}, concentrationCurrencyID = 0, concentrationCost = 0, ingenuityRefund = 0 }
+    end
+
     local res
     if type(orderOrRecraftGUID) == "table" then
+        ---@todo DEBUG Some player orders cause freezes
+        -- if orderOrRecraftGUID.orderType ~= Enum.CraftingOrderType.Npc then
+        --     n = n + 1
+        --     local i = n
+        --     local form = Addon.GUI.RecipeForm.OrdersForm.form
+
+        --     local origReagents = reagents
+        --     local formReagents = form and form.transaction and form:IsVisible() and form.transaction:CreateCraftingReagentInfoTbl()
+
+        --     Addon:Debug(origReagents, "reagents " .. i .. " (orig)")
+        --     Addon:Debug(formReagents, "reagents " .. i .. " (form)")
+
+        --     reagents = formReagents or {}
+
+        --     C_Timer.After(0, function ()
+        --         Addon:Debug(true, "> " .. i)
+        --         local op = C_TradeSkillUI.GetCraftingOperationInfoForOrder(recipe.recipeID, origReagents, orderOrRecraftGUID.orderID, applyConcentration)
+        --         Addon:Debug(op, "> op " .. i)
+        --     end)
+        -- end
+
         res = C_TradeSkillUI.GetCraftingOperationInfoForOrder(recipe.recipeID, reagents, orderOrRecraftGUID.orderID, applyConcentration)
     else ---@cast orderOrRecraftGUID string?
         res = C_TradeSkillUI.GetCraftingOperationInfo(recipe.recipeID, reagents, orderOrRecraftGUID, applyConcentration)